+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.10.0] - 2026-04-30
 ### Added
 - Added `ProjectService` with `GetProject` method to retrieve a project data
 - Added `GetProjectByPurlName(purlName, purlType)` method in `ProjectModel` to look up a single project by PURL name and type
@@ -118,3 +120,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.7.0]: https://github.com/scanoss/go-models/compare/v0.6.0...v0.7.0
 [0.8.0]: https://github.com/scanoss/go-models/compare/v0.7.0...v0.8.0
 [0.9.0]: https://github.com/scanoss/go-models/compare/v0.8.0...v0.9.0
+[0.10.0]: https://github.com/scanoss/go-models/compare/v0.9.0...v0.10.0